Move test_network_group into the TestProtocol
